Rename view package into output
