+++ cli/workflowCommands.go
@@ -57,7 +57,7 @@ import (
 	"github.com/temporalio/tctl/common/payload"
 	"github.com/temporalio/tctl/common/payloads"
 	"github.com/temporalio/tctl/pkg/color"
-	"github.com/temporalio/tctl/pkg/view"
+	"github.com/temporalio/tctl/pkg/output"
 	clispb "go.temporal.io/server/api/cli/v1"
 	"go.temporal.io/server/common/clock"
 	"go.temporal.io/server/common/convert"
@@ -111,8 +111,8 @@ func ShowHistory(c *cli.Context) {
 	}
 
 	iter := collection.NewPagingIterator(paginationFunc)
-	opts := &view.PrintOptions{Fields: []string{"ID", "Type", "Event"}}
-	view.Paginate(c, iter, opts)
+	opts := &output.PrintOptions{Fields: []string{"ID", "Type", "Event"}}
+	output.Pager(c, iter, opts)
 }
 
 // RunWorkflow starts a new workflow execution and print workflow progress and result
@@ -188,8 +188,8 @@ func RunWorkflow(c *cli.Context) {
 		&row{Field: "Args", Value: truncate(payloads.ToString(input))},
 	}
 	fmt.Println(color.Magenta(c, "Running execution:"))
-	opts := &view.PrintOptions{Fields: []string{"Field", "Value"}, NoHeader: true}
-	view.PrintItems(c, executionData, opts)
+	opts := &output.PrintOptions{Fields: []string{"Field", "Value"}, NoHeader: true}
+	output.PrintItems(c, executionData, opts)
 
 	printWorkflowProgress(c, wid, resp.GetRunId())
 }
@@ -316,7 +316,7 @@ func printWorkflowProgress(c *cli.Context, wid, rid string) {
 	timeElapse := 1
 	isTimeElapseExist := false
 	ticker := time.NewTicker(time.Second).C
-	opts := &view.PrintOptions{
+	opts := &output.PrintOptions{
 		Fields:     []string{"ID", "Time", "Type"},
 		FieldsLong: []string{"Details"},
 		All:        true,
@@ -327,7 +327,7 @@ func printWorkflowProgress(c *cli.Context, wid, rid string) {
 	go func() {
 		hIter := sdkClient.GetWorkflowHistory(tcCtx, wid, rid, true, enumspb.HISTORY_EVENT_FILTER_TYPE_ALL_EVENT)
 		iter := &historyIterator{iter: hIter, maxFieldLength: maxFieldLength, lastEvent: &lastEvent}
-		view.Paginate(c, iter, opts)
+		output.Pager(c, iter, opts)
 
 		doneChan <- true
 	}()
@@ -516,11 +516,11 @@ func ListWorkflow(c *cli.Context) {
 	}
 
 	iter := collection.NewPagingIterator(paginationFunc)
-	opts := &view.PrintOptions{
+	opts := &output.PrintOptions{
 		Fields:     []string{"Execution.WorkflowId", "Execution.RunId", "StartTime"},
 		FieldsLong: []string{"Type.Name", "TaskQueue", "ExecutionTime", "CloseTime"},
 	}
-	view.Paginate(c, iter, opts)
+	output.Pager(c, iter, opts)
 }
 
 // ScanAllWorkflow list all workflow executions using Scan API.
@@ -556,11 +556,11 @@ func ScanAllWorkflow(c *cli.Context) {
 	}
 
 	iter := collection.NewPagingIterator(paginationFunc)
-	opts := &view.PrintOptions{
+	opts := &output.PrintOptions{
 		Fields:     []string{"Execution.WorkflowId", "Execution.RunId", "StartTime"},
 		FieldsLong: []string{"Type.Name", "TaskQueue", "ExecutionTime", "CloseTime"},
 	}
-	view.Paginate(c, iter, opts)
+	output.Pager(c, iter, opts)
 }
 
 // CountWorkflow count number of workflows
@@ -621,11 +621,11 @@ func ListArchivedWorkflow(c *cli.Context) {
 	}
 
 	iter := collection.NewPagingIterator(paginationFunc)
-	opts := &view.PrintOptions{
+	opts := &output.PrintOptions{
 		Fields:     []string{"Execution.WorkflowId", "Execution.RunId", "StartTime"},
 		FieldsLong: []string{"Type.Name", "TaskQueue", "ExecutionTime", "CloseTime"},
 	}
-	view.Paginate(c, iter, opts)
+	output.Pager(c, iter, opts)
 }
 
 // DescribeWorkflow show information about the specified workflow execution
