+++ pkg/output/json.go
@@ -22,47 +22,29 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package view
+package output
 
 import (
-	"github.com/olekukonko/tablewriter"
-	"github.com/urfave/cli/v2"
+	"encoding/json"
+	"fmt"
 
-	"github.com/temporalio/tctl/pkg/process"
+	"github.com/gogo/protobuf/proto"
+	"github.com/temporalio/shared-go/codec"
 )
 
-var (
-	headerColor = tablewriter.Colors{tablewriter.FgHiBlueColor}
-)
-
-func PrintTable(c *cli.Context, items []interface{}, opts *PrintOptions) {
-	fields := opts.Fields
-	table := tablewriter.NewWriter(opts.Pager)
-	table.SetBorder(false)
-	table.SetColumnSeparator(opts.Separator)
-
-	if !opts.NoHeader {
-		table.SetHeader(fields)
-		headerColors := make([]tablewriter.Colors, len(fields))
-		for i := range headerColors {
-			headerColors[i] = headerColor
-		}
-		table.SetHeaderColor(headerColors...)
-		table.SetHeaderLine(false)
+func PrintJSON(o interface{}, opts *PrintOptions) {
+	var b []byte
+	var err error
+	if pb, ok := o.(proto.Message); ok {
+		encoder := codec.NewJSONPBIndentEncoder("  ")
+		b, err = encoder.Encode(pb)
+	} else {
+		b, err = json.MarshalIndent(o, "", "  ")
 	}
 
-	rows, err := extractFieldValues(items, fields)
 	if err != nil {
-		process.ErrorAndExit("unable to print table", err)
+		fmt.Printf("Error when try to print pretty: %v\n", err)
+		fmt.Fprintln(opts.Pager, o)
 	}
-
-	for _, row := range rows {
-		columns := make([]string, len(row))
-		for j, column := range row {
-			columns[j] = formatField(c, column)
-		}
-		table.Append(columns)
-	}
-	table.Render()
-	table.ClearRows()
+	fmt.Fprintln(opts.Pager, string(b))
 }
